add position functionnal in SFMLGraphic.cpp

+++ R-Type/client/includes/GUI/SFMLGraphic.hpp
@@ -24,9 +24,9 @@ class SFMLGraphic : public IGraphic
 
 	// public methods
 	public:
-		bool	drawSprite(std::string const &key, float delta, uint32_t x, uint32_t y);
+		bool	drawSprite(std::string const &key, float delta, float x, float y);
 		void	update();
-		bool	drawFont(std::string const &key, std::string const &str, uint32_t x, uint32_t y);
+		bool	drawFont(std::string const &key, std::string const &str, float x, float y);
 		bool	playSound(bool onLoop = false);
 		bool	isOpen() const;
 		void	handleEvent();
